+++ Ajuna.NetApi.Test/Extrinsic/EraTest.cs
@@ -1,6 +1,5 @@
 using System;
 using NUnit.Framework;
-using Ajuna.NetApi;
 using Ajuna.NetApi.Model.Extrinsics;
 
 namespace Ajuna.NetApi.Test.Extrinsic
