+++ Ajuna.NetApi.Test/Extrinsic/SignedExtensionsTest.cs
@@ -1,6 +1,5 @@
 using System;
 using NUnit.Framework;
-using Ajuna.NetApi;
 using Ajuna.NetApi.Model.Extrinsics;
 using Ajuna.NetApi.Model.Types.Base;
