+++ Ajuna.NetApi.Test/Keys/Ed25519Tests.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq;
 using NUnit.Framework;
-using Ajuna.NetApi;
 using Ajuna.NetApi.Model.Types.Base;
 
 namespace Ajuna.NetApi.Test.Keys
