+++ Ajuna.NetApi.Test/RequestGeneratorTests.cs
@@ -1,5 +1,4 @@
 using NUnit.Framework;
-using Ajuna.NetApi;
 using Ajuna.NetApi.Model.Meta;
 
 namespace Ajuna.NetApi.Test
@@ -16,6 +15,13 @@ namespace Ajuna.NetApi.Test
         {
             var result = RequestGenerator.GetStorage("Sudo", "Key", Storage.Type.Plain);
             Assert.AreEqual("0x5C0D1176A568C1F92944340DBFED9E9C530EBCA703C85910E7164CB7D1C9E47B", result);
+        }
+
+        [Test]
+        public void GetKeyBytesTest()
+        {
+            var result = Utils.Bytes2HexString(RequestGenerator.GetStorageKeyBytesHash("System", "Account"));
+            Assert.AreEqual("0x26AA394EEA5630E07C48AE0C9558CEF7B99D880EC681799C0CF30E8886371DA9", result);
         }
     }
 }
