+++ Ajuna.NetApi.Test/UtilsTests.cs
@@ -1,6 +1,5 @@
 using System;
 using NUnit.Framework;
-using Ajuna.NetApi;
 
 namespace Ajuna.NetApi.Test
 {
