+++ Ajuna.NetApi/Ajuna.NetApi.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <PackageId>Ajuna.NetApi</PackageId>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <Version>0.1.2</Version>
+    <Version>0.1.3-pre.0</Version>
     <Company>BloGa Tech AG</Company>
     <Authors>Cedric Decoster</Authors>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
