--- Ajuna.NetApi/Model/Custom/Template/TemplateCalls.cs
@@ -1,7 +1,0 @@
-﻿namespace Ajuna.NetApi.Model.Calls
-{
-    public class TemplateCall
-    {
-
-    }
-}
