--- Ajuna.NetApi/Model/Custom/Template/TemplateTypes.cs
@@ -1,10 +1,0 @@
-﻿using Ajuna.NetApi.Model.Types.Primitive;
-
-namespace Ajuna.NetApi.Model.Types.Custom
-{
-    public class CustomU32 : U32
-    {
-        public override string TypeName() => "CustomU32";
-    }
-
-}
