--- Ajuna.NetApi/Model/Types/Enum/NetApiTestEnum.cs
@@ -1,7 +1,0 @@
-﻿namespace Ajuna.NetApi.Model.Types.Enum
-{
-    public enum NetApiTestEnum
-    {
-        None
-    }
-}
