+++ Ajuna.NetApi/Model/Types/Metadata/V14/PortableRegistry.cs
