+++ Ajuna.NetApi/Model/Types/Metadata/V14/Variant.cs
@@ -1,6 +1,5 @@
 ﻿using Ajuna.NetApi.Model.Types.Base;
 using Ajuna.NetApi.Model.Types.Primitive;
-using Ajuna.NetApi.Model.Types.Struct;
 using System;
 
 namespace Ajuna.NetApi.Model.Types.Metadata.V14
