--- Ajuna.NetApi/Model/Types/Struct/AccountData.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using Ajuna.NetApi.Model.Types.Base;
-
-namespace Ajuna.NetApi.Model.Types.Struct
-{
-    public class AccountData : BaseType
-    { 
-        public override string TypeName() => "AccountData<T::Balance>";
-                
-        public override byte[] Encode()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void Decode(byte[] byteArray, ref int p)
-        {
-            var start = p;
-
-            Free = new Balance();
-            Free.Decode(byteArray, ref p);
-
-            Reserved = new Balance();
-            Reserved.Decode(byteArray, ref p);
-
-            MiscFrozen = new Balance();
-            MiscFrozen.Decode(byteArray, ref p);
-
-            FeeFrozen = new Balance();
-            FeeFrozen.Decode(byteArray, ref p);
-
-            TypeSize = p - start;
-        }
-
-        public Balance Free { get; private set; }
-        public Balance Reserved { get; private set; }
-        public Balance MiscFrozen { get; private set; }
-        public Balance FeeFrozen { get; private set; }
-    }
-}
