--- Ajuna.NetApi/Model/Types/Struct/AccountInfo.cs
@@ -1,40 +1,0 @@
-﻿using System;
-using Ajuna.NetApi.Model.Types.Base;
-using Ajuna.NetApi.Model.Types.Primitive;
-
-namespace Ajuna.NetApi.Model.Types.Struct
-{
-    public class AccountInfo : BaseType
-    {
-        public override string TypeName() => "AccountInfo<T::Index, T::AccountData>";
-
-        public override byte[] Encode()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void Decode(byte[] byteArray, ref int p)
-        {
-            var start = p;
-
-            Nonce = new U32();
-            Nonce.Decode(byteArray, ref p);
-
-            Consumers = new RefCount();
-            Consumers.Decode(byteArray, ref p);
-
-            Providers = new RefCount();
-            Providers.Decode(byteArray, ref p);
-
-            AccountData = new AccountData();
-            AccountData.Decode(byteArray, ref p);
-
-            TypeSize = p - start;
-        }
-
-        public U32 Nonce { get; private set; }
-        public RefCount Consumers { get; private set; }
-        public RefCount Providers { get; private set; }
-        public AccountData AccountData { get; private set; }
-    }
-}
