--- Ajuna.NetApi/Model/Types/UnimplementedTypes.cs
@@ -1,37 +1,0 @@
-﻿using System;
-
-namespace Ajuna.NetApi.Model.Types
-{
-    public class DispatchError
-    {
-        public static DispatchError Decode(Memory<byte> byteArray, ref int p)
-        {
-            throw new NotImplementedException("DispatchError");
-        }
-    }
-
-    public class DispatchResult
-    {
-        public static DispatchResult Decode(Memory<byte> byteArray, ref int p)
-        {
-            throw new NotImplementedException("DispatchResult");
-        }
-    }
-
-    public class Kind
-    {
-        public static Kind Decode(Memory<byte> byteArray, ref int p)
-        {
-            throw new NotImplementedException("Kind");
-        }
-    }
-
-    public class Moment
-    {
-        public byte[] Encode()
-        {
-            throw new NotImplementedException("Moment");
-        }
-    }
-    
-}
